requirements.txt
streamlit==1.28.0
openai==1.3.0
chromadb==0.4.22
sqlite-utils==3.36
docker==6.1.3
python-dotenv==1.0.0
pydantic==2.4.0
pyjwt==2.8.0
